feat(auth): Auth with signout
