--- frontend/src/index.ts
@@ -1,11 +1,0 @@
-// import 'dotenv/config';
-// import { drizzle } from 'drizzle-orm/postgres-js'
-// import { eq } from 'drizzle-orm';
-// import { users } from './db/schema'
-
-// const connectionString = process.env.VITE_DATABASE_URL;
-
-// const db = drizzle(connectionString!); //Disable prefetch as it is not supported for "Transaction" pool mode
-
-// const allUsers = await db.select().from(users);
-// console.log(allUsers)
