+++ frontend/src/main.tsx
@@ -1,34 +1,10 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
-import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import "./index.css";
 import App from "./App.tsx";
-import Login from "./pages/Login.tsx";
-import Signup from "./pages/Signup.tsx";
-import ForgotPassword from "./pages/ForgotPassword.tsx";
-import UpdatePassword from "./pages/UpdatePassword.tsx";
-import AuthenticatedRoute from "./pages/AuthentiatedRoute.tsx";
-import { AuthProvider } from "./context/AuthContext.tsx";
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
-    <AuthProvider>
-      <Router>
-        <Routes>
-          <Route
-            path="/"
-            element={
-              <AuthenticatedRoute>
-                <App />
-              </AuthenticatedRoute>
-            }
-          />
-          <Route path="/login" element={<Login />} />
-          <Route path="/sign-up" element={<Signup />} />
-          <Route path="/forgot-password" element={<ForgotPassword />} />
-          <Route path="/update-password" element={<UpdatePassword />} />
-        </Routes>
-      </Router>
-    </AuthProvider>
+    <App />
   </StrictMode>
 );
