+++ frontend/src/pages/ForgotPassword.tsx
@@ -1,8 +1,6 @@
 import { ForgotPasswordForm } from "@/components/forgot-password-form";
-import { useSession } from "@/context/AuthContext";
 
 function ForgotPassword() {
-  const session = useSession();
   return (
     <>
       <ForgotPasswordForm />
