+++ frontend/src/pages/Login.tsx
@@ -1,8 +1,6 @@
 import { LoginForm } from "@/components/login-form";
-import { useSession } from "@/context/AuthContext";
 
 function Login() {
-  const session = useSession();
   return (
     <>
       <LoginForm />
