+++ frontend/src/pages/Signup.tsx
@@ -1,8 +1,6 @@
 import { SignUpForm } from "@/components/sign-up-form";
-import { useSession } from "@/context/AuthContext";
 
 function Signup() {
-  const session = useSession();
   return (
     <>
       <SignUpForm />
