+++ frontend/src/pages/UpdatePassword.tsx
@@ -1,8 +1,6 @@
 import { UpdatePasswordForm } from "@/components/update-password-form";
-import { useSession } from "@/context/AuthContext";
 
 function UpdatePassword() {
-  const session = useSession();
   return (
     <>
       <UpdatePasswordForm />
